backend/requirements.txt
# Flask and web framework
Flask==3.0.0
flask-cors==4.0.0

# AI and ML
requests==2.31.0

# Document processing
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.22
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
PyMuPDF==1.23.8

# Data processing
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
werkzeug==3.0.1
python-dotenv==1.0.0